pyproject.toml
[tool.poetry]
name = "thecapitalfund"
version = "0.1.0"
description = "The Capital Fund"
authors = ["Tom <t.stalley@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
dash = "2.13.0"
dash-bootstrap-components = "1.4.2"
pandas = "2.0.3"
plotly = "5.16.1"
pydantic = "2.3.0"
requests = "2.31.0"
argon2-cffi = "^23.1.0"
numpy = "^1.24.0"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
ruff = "^0.8.0"

[tool.poetry.group.dev.dependencies]
deptry = "^0.16.1"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
